chore: run benchmarks in separate VMs
